+++ src/components/SlackErrorDisplay.tsx
@@ -1,0 +1,37 @@
+/**
+ * Slackエラー表示コンポーネント
+ * - エラーメッセージの表示
+ * - エラーアイコンとメッセージのスタイリング
+ */
+
+type SlackErrorDisplayProps = {
+  error: string | null
+}
+
+export function SlackErrorDisplay({ error }: SlackErrorDisplayProps) {
+  if (!error) return null
+
+  return (
+    <div className="mt-5 p-3.5 text-sm text-gray-800 bg-red-50 border border-red-300 rounded-sm shadow-sm">
+      <div className="flex items-center">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          className="h-5 w-5 mr-2 text-red-500"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+          strokeWidth="2"
+        >
+          <title>エラーアイコン</title>
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
+          />
+        </svg>
+        <p className="font-medium">エラーが発生しました:</p>
+      </div>
+      <p className="ml-7 mt-0.5 text-red-600">{error}</p>
+    </div>
+  )
+}
